fixes for MacOS X
git-svn-id: https://bugengine.googlecode.com/svn/trunk@361 f3700e90-1ba7-11de-b56b-c9554f25afe1

+++ src/engine/rtti/src/typeinfo.cc
@@ -8,7 +8,7 @@
 namespace BugEngine
 {
 
-size_t TypeInfo::size() const
+u32 TypeInfo::size() const
 {
     switch(type)
     {
@@ -150,7 +150,7 @@ static struct A
         v("name") = c;
         OutputDebugString((::minitl::format<>("name: %s\n") | name.as< raw<const char> >().m_ptr).c_str());
         OutputDebugString((::minitl::format<>("name: %s\n") | ci.name.m_ptr).c_str());
-        
+
         RTTI::ClassInfo root = *be_typeid<void>().klass;
     }
     ~A()
